enable warning as errors

+++ symget.c
@@ -349,7 +349,7 @@ static DWORD WINAPI ProcessThread(LPVOID arg)
 					if (base)
 					{
 						const IMAGE_DOS_HEADER* dos = base;
-						if (dos->e_magic == IMAGE_DOS_SIGNATURE && ((SIZE_T)dos->e_lfanew + sizeof(IMAGE_NT_HEADERS) < size.QuadPart))
+						if (dos->e_magic == IMAGE_DOS_SIGNATURE && ((SIZE_T)dos->e_lfanew + sizeof(IMAGE_NT_HEADERS) < (SIZE_T)size.QuadPart))
 						{
 							const IMAGE_NT_HEADERS* nt = (void*)((char*)base + dos->e_lfanew);
 							if (nt->Signature == IMAGE_NT_SIGNATURE)
